NexmoServiceProvider.php
Thanks so much for this great tool! This error stumped me for a good while. The way I was reading the error I thought it wasn't getting either the api_secret or the signature_secret. I knew I was providing the api_secret so I wasn't understanding what the problem could be. After a bit of head scratching and digging I found this line in the code and realized that I was providing both the api_secret and the signature_secret values in my .env file, I didn't know that would cause a problem. When I removed the signature_secret value the code worked perfectly. In conclusion, this is just an updated error message to help the next person understand what they're doing wrong. Thanks again.